% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchangeableSampling.R
\name{exchangeableSampling}
\alias{exchangeableSampling}
\title{Exchangeable Resampling for GGMs}
\usage{
exchangeableSampling(
  data,
  adjacency,
  target = 1:ncol(data),
  m = 500,
  l = 3,
  ncores = 1
)
}
\arguments{
\item{data}{A data.frame or matrix containing all variables to include in the
network.}

\item{adjacency}{An adjacency matrix indicating whether an edge is assumed (1)
or not assumed (0) between two variables. Rows and columns must be in the
same order as the columns of \code{data}.}

\item{target}{The columns of \code{data} which are nodes of the target system.}

\item{m}{The number of exchangeable copy datasets to generate.}

\item{l}{The number of passes of the markov chain.}

\item{ncores}{The number of cores used in parallel computing.}
}
\value{
A list containing the (standardized) original dataset, the hub data, and a list of
all generated copies.
}
\description{
This is a function to generate exchangeable copies of a dataset based on the
exchangeable sampling algorithm propose by Lin et al. (2025).
}
\references{
Lin, X., Li, W., Tian, F., & Huang, D. (2025). Goodness-of-fit
tests for high-dimensional Gaussian graphical models via exchangeable
sampling. Journal of the Royal Statistical Society Series B:
Statistical Methodology, qkaf032. https://doi.org/10.1093/jrsssb/qkaf032
}
