% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametricBootstrap.R
\name{parametricBootstrap}
\alias{parametricBootstrap}
\title{Parametric Bootstrap for GGMs}
\usage{
parametricBootstrap(data, adjacency, m = 500, ncores = 1)
}
\arguments{
\item{data}{A data.frame or matrix containing all variables to include in the
network.}

\item{adjacency}{An adjacency matrix indicating whether an edge is assumed (1)
or not assumed (0) between two variables. Rows and columns must be in the
same order as the columns of \code{data}.}

\item{m}{The number of bootstrap samples to generate.}

\item{ncores}{The number of cores used in parallel computing.}
}
\value{
A list containing the (standardized) original dataset and a list of
all generated copies.
}
\description{
This is a function to generate parametric bootstrap samples of a dataset,
given a proposed adjacency matrix.
}
