% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{plot.hyponetResult}
\alias{plot.hyponetResult}
\title{Plot the results of a network hypothesis test}
\usage{
\method{plot}{hyponetResult}(x, which = "all", ask = FALSE, breaks = "FD", mfrow = NULL, ...)
}
\arguments{
\item{x}{An object of class hyponetResult}

\item{which}{Which of the fit statistics to plot. Either 'all' (the default),
a vector of names of the fit statistics, or a numeric vector indicating which
of the fit values provided in the hyponetResult object to plot.}

\item{ask}{A logical indicating whether to ask before drawing the next plot
(if more than one fit statistic is being plotted).}

\item{breaks}{A breaks argument as described in \code{?hist}. Defaults to
'FD' for the use of \code{nclass.FD}.}

\item{mfrow}{A vector of two numbers setting the number of rows and columns
to be displayed when plotting multiple fit statistics.}

\item{...}{Other arguments to be passed to \code{hist}.}
}
\description{
Plot the results of a network hypothesis test
}
