% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{print.hyponetResult}
\alias{print.hyponetResult}
\title{Print an object of class hyponetResult}
\usage{
\method{print}{hyponetResult}(x, ...)
}
\arguments{
\item{x}{An object of class hyponetResult}

\item{...}{Currently not used}
}
\description{
Print an object of class hyponetResult
}
