% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{print.summary.hyponetResult}
\alias{print.summary.hyponetResult}
\title{Print the summary of a hyponetResult}
\usage{
\method{print}{summary.hyponetResult}(x, ...)
}
\arguments{
\item{x}{An object of class summary.hyponetResult}

\item{...}{Currently not used}
}
\description{
Print the summary of a hyponetResult
}
