% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{summary.hyponetResult}
\alias{summary.hyponetResult}
\title{Summarize an object of class hyponetResult}
\usage{
\method{summary}{hyponetResult}(object, ...)
}
\arguments{
\item{object}{An object of class hyponetResult}

\item{...}{Currently not used}
}
\description{
Summarize an object of class hyponetResult
}
